Move download code to new `brioche_core::download` module
